+++ src/providers/LocalVideoProvider/index.tsx
@@ -23,7 +23,7 @@ const LocalVideoProvider: React.FC = ({ children }) => {
   const logger = useLogger();
   const meetingManager = useMeetingManager();
   const audioVideo = useAudioVideo();
-  const { selectedDevice } = useVideoInputs();
+  const { devices, selectedDevice } = useVideoInputs();
   const [isVideoEnabled, setIsVideoEnabled] = useState(false);
   const [hasReachedVideoLimit, setHasReachedVideoLimit] = useState(false);
   const [tileId, setTileId] = useState<number | null>(null);
@@ -62,6 +62,14 @@ const LocalVideoProvider: React.FC = ({ children }) => {
       logger.warn('Reach the number of maximum active videos');
     }
   }, [hasReachedVideoLimit]);
+
+  // In the case that the selected device is unplugged, the JS SDK will automatically call stopLocalVideoTile
+  // We can then set the isVideoEnabled to false
+  useEffect(() => {
+    if (!audioVideo?.hasStartedLocalVideoTile()) {
+      setIsVideoEnabled(false);
+    }
+  }, [devices]);
 
   const toggleVideo = useCallback(async (): Promise<void> => {
     try {